move the common value to monitor
