+++ rmm/armv9a/src/realm/mod.rs
@@ -1,4 +1,4 @@
 pub mod context;
 pub mod mm;
-pub mod registry;
 pub mod vcpu;
+pub mod vm;
