+++ rmm/armv9a/src/rmi/version.rs
@@ -1,9 +1,9 @@
-use monitor::{listen, mainloop::Mainloop};
+use monitor::{listen, mainloop::Mainloop, rmi};
 
 use crate::config;
-use crate::rmi;
+use crate::rmi::Receiver;
 
-pub fn set_event_handler(mainloop: &mut Mainloop<rmi::Receiver>) {
+pub fn set_event_handler(mainloop: &mut Mainloop<Receiver>) {
     listen!(mainloop, rmi::Code::Version, |call| {
         trace!("RMM: requested version information");
         Ok(call.reply(config::ABI_VERSION)?)
